front end files
chat.js contains functions that can parse, receive, and display data. The register page also contains a field to upload an image for a profile picture. This is base front end code, and once functionality for different features are made, I can work on making the pages look very nice.